Fixed Laba 2
