+++ test8/src/Solution.java
@@ -1,72 +1,22 @@
-﻿import java.io.*;
-import java.util.*;
-import java.math.*;
+import java.util.ArrayList;
+import java.util.Collections;
+
 
 public class Solution {
-   /*
-    * Определить максимальную мощность подмножества элементов заданного массива. Элементы включаются в подмножество, если
-	* разность между любыми элементами подмножества меньше либо равна 1.	
-	* Дано: n - количество чисел в массиве
-	*		a - массив чисел
-	* 		2 <= n <= 100
-	*		0 < a[i] < 100
-	*		ответ должен быть >=2, задаются такие данные
-	* Пример:
-	*  9
-	*  1 1 2 2 4 4 5 5 5
-	* Ответ:
-	* 5
-	*
-	* 6
-	* 4 6 5 3 3 1
-	* Ответ:
-	* 3
-	*/
-	
-    static int func1(int[] a) {
-        Arrays.sort(a);
-        int max = 0, 
-			pmax = 1,
-			i = 0,
-			j = 1;
-       
-	   while (i < a.length) {
+    static int func1(ArrayList a) {
+        Collections.sort(a);
+        int max = 0, pmax = 1,i = 0,j = 1;
+        while (i < a.size()) {
             if (max < pmax)
                 max = pmax;
             pmax = 1;
-            while (j < a.length && Math.abs(a[i] - a[j]) <= 1) {
+            while (j < a.size() && Math.abs((int)a.get(i) - (int)a.get(j)) <= 1) {
                 pmax += 1;
                 j += 1;
             }
-            i ++;;
+            i += 1;
             j = i + 1;
         }
         return max;
-    }
-
-    private static final Scanner scanner = new Scanner(System.in);
-
-    public static void main(String[] args) throws IOException {
-        BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter("out.txt"));//System.getenv("OUTPUT_PATH")
-
-        int n = scanner.nextInt();
-        scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");
-
-        int[] a = new int[n];
-
-        String[] aItems = scanner.nextLine().split(" ");
-        scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");
-
-        for (int i = 0; i < n; i++) {
-            int aItem = Integer.parseInt(aItems[i]);
-            a[i] = aItem;
-        }
-
-        int result = func1(a);
-	
-	    bufferedWriter.write(String.valueOf(result));
-       
-        bufferedWriter.close();
-        scanner.close();
     }
 }
